Adds example code for individual finger control
